fix database initialization
